move role event handler back to workloads controller
